Added Playlist Sync Functionality
Now, the user can access their playlist from anywhere. On startup the Client
Desktop Application will notify the server about what playlists and songs
the client has. The server will create a diff that includes the changes the client
needs to do. The client takes the diff and updates the playlists and the songs
accordingly.