EventSystem in Client Desktop Application has been updated to support Event Args.
Before this version, we were casting arguments that
were supplied by the event. If the arguments were
not ordered correctly, or they were not appropriate, the method
handlers would throw an Invalid Cast Exception.
In this version we can specify the type of Arguments when we Fire
and Event. Also, this event will only be handled if the correct Argument type
is specified in the Delegate in the Listener.