Added 3 properties in DataAccessConfigurationService that can dynamically update appsettings.
1. DesktopAppSongSearchLimit: this property will limit the number of songs displayed
when the client searches for songs in the desktop application. This will make sure that
the network and memory are not exhausted when the search is yet not specific.

2. WebAppSongSearchLimit: similar to DesktopAppSongSearchLimit, this limits the number
of songs displayed. However, the change is for the Administrator role in the web interface.

3. DesktopAppClientCountLimit: this property limits the number of desktop client
that can be connected to the Audio Engine.

All these properties can dynamically change appsettings and IConfiguration. builder.Configuration
has "reloadOnChange" set to true. This will make sure that when the Audio Engine User changes the
settings, the settings are immediately applied.