Fixed a bug in the Server/Controller.
When a desktop client would try to connect while the Controller
had reached client limit, the controller would fail if the desktop
client closed their applcation, and then tried to reconnect.

This was caused by "socket.Accept()" since it is a blocking operation.
The thread would most of the time be waiting for "socket.Accept()" before
checking the client limit.